+++ app/Providers/EventServiceProvider.php
@@ -2,35 +2,21 @@
 
 namespace App\Providers;
 
-use App\Events\LinkArchiveRequested;
-use App\Listeners\MakeLinkArchive;
 use Illuminate\Support\Facades\Event;
-use Illuminate\Auth\Events\Registered;
-use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
 
 class EventServiceProvider extends ServiceProvider
 {
-    /**
-     * The event listener mappings for the application.
-     *
-     * @var array
-     */
     protected $listen = [
-        LinkArchiveRequested::class => [
-            MakeLinkArchive::class
+        \App\Events\LinkArchiveRequested::class => [
+            \App\Listeners\MakeLinkArchive::class
         ]
     ];
 
-    /**
-     * Register any events for your application.
-     *
-     * @return void
-     */
     public function boot()
     {
         parent::boot();
 
-        //
+        \App\Tag::observe(\App\Observers\TagObserver::class);
     }
 }